made an component of "tredning movies" to make the code easier to follow and read
